Migrate all shader-const modification classes out of header files
Each of these classes is essentially a single function (a functor), so
giving them their own files wasn't necessary. And having them in source
rather than headers enables experimentation without a full recompile.

Shader-modifications that were unused (MatrixHudSquash,
MatrixNoShowModification) are now wrapped in "#if 0".
